chore: refactor conftest and extract constants
