+++ tests/tests_utils/constants.py
@@ -1,0 +1,8 @@
+EXIT_PERIOD = 4 * 60  # 4 minutes
+INITIAL_IMMUNE_VAULTS = 1
+INITIAL_IMMUNE_EXIT_GAMES = 1
+
+GAS_LIMIT = 10000000
+START_GAS = GAS_LIMIT - 1000000
+
+HUNDRED_ETH = 100 * 10 ** 18
